feat: split manifest up into smaller files + automatically retain existing behavior if issues are found.
